RFC-0020/0021, strategy docs, CAC schemas, security migration, and code fixes (#430)
* RFC-0020/0021, strategy docs, CAC schemas, security migration, and code fixes

- Add RFC-0020 Holonic Substrate Interface (full spec with 36 requirements)
- Add RFC-0021 Holonic Venture Proving Interface with evidence artifacts
- Add business strategy protocol instructions, reports, and exercises
- Add CAC schemas (attention_decision, venture_signal, strategy_coupling_ledger, etc.)
- Migrate security docs from markdown to CAC JSON format
- Add JSON schemas for security document types
- Update strategy documents (BUSINESS_PLAN, MASTER_STRATEGY, ROADMAP, NORTH_STAR)
- Add work tickets TCK-00347 through TCK-00382
- Fix CLI coordinate command, seccomp adapter, PTY episode, proc_fallback telemetry
- Update proto kernel_events, xtask reviewer_state and related tooling


* Fix CI gate failures and include branch updates

---------

+++ crates/apm2-daemon/src/episode/pty.rs
@@ -519,7 +519,12 @@ impl PtyRunner {
                 // Set slave as controlling terminal
                 // SAFETY: TIOCSCTTY is a valid ioctl for setting controlling terminal
                 unsafe {
-                    if libc::ioctl(pty.slave.as_raw_fd(), libc::TIOCSCTTY, 0) < 0 {
+                    if libc::ioctl(
+                        pty.slave.as_raw_fd(),
+                        libc::c_ulong::from(libc::TIOCSCTTY),
+                        0,
+                    ) < 0
+                    {
                         // Best effort - continue anyway (some systems don't
                         // require this)
                     }
